+++ core/src/main/java/com/principal/AntsArmageddon.java
@@ -5,6 +5,7 @@ import com.badlogic.gdx.assets.AssetManager;
 import managers.GestorAssets;
 import managers.ScreenManager;
 import screens.MenuScreen;
+import utils.RecursosGlobales;
 
 /** {@link com.badlogic.gdx.ApplicationListener} implementation shared by all platforms. */
 public class AntsArmageddon extends Game {
@@ -13,6 +14,7 @@ public class AntsArmageddon extends Game {
 
     @Override
     public void create() {
+        RecursosGlobales.inicializar();
         GestorAssets.load();
 
         ScreenManager.setJuego(this);
@@ -22,5 +24,5 @@ public class AntsArmageddon extends Game {
     @Override
     public void render() { super.render(); }
     @Override
-    public void dispose() { super.dispose(); GestorAssets.dispose(); }
+    public void dispose() { super.dispose(); GestorAssets.dispose(); RecursosGlobales.dispose(); }
 }