+++ core/src/main/java/hud/Hud.java
@@ -4,31 +4,46 @@ import Fisicas.Camara;
 import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.graphics.g2d.BitmapFont;
 import com.badlogic.gdx.graphics.g2d.GlyphLayout;
-import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.utils.viewport.Viewport;
 import entidades.personajes.Personaje;
+import managers.GestorAssets;
+import utils.RecursosGlobales;
 
 public class Hud {
-    private final BitmapFont fuente; //cambiar la fuente con un .fnt
+    private final BitmapFont fuenteContador, fuenteVida;
     private final GlyphLayout layout;
 
     public Hud() {
-        this.fuente = new BitmapFont();
+        this.fuenteContador = GestorAssets.get("fonts/font_Contador.fnt", BitmapFont.class);
+        this.fuenteVida = GestorAssets.get("fonts/font_Vida.fnt", BitmapFont.class);
         this.layout = new GlyphLayout();
     }
 
-    public void mostrarVida(SpriteBatch batch, Personaje personaje) {
+    public void mostrarVida(Personaje personaje) {
         if (!personaje.getActivo()) return;
-        String texto = personaje.getVida() + " / " + personaje.getVidaMaxima();
-        layout.setText(fuente, texto);
-        float posX = personaje.getX();
+
+        String texto = String.valueOf(personaje.getVida());
+        layout.setText(fuenteVida, texto);
+
+        float posX = personaje.getX() + personaje.getSprite().getWidth() / 2f - layout.width / 2f;
         float posY = personaje.getY() + personaje.getSprite().getHeight() + 20;
-        fuente.draw(batch, texto, posX, posY);
+
+        float offset = 1f;
+
+        fuenteVida.setColor(0, 0, 0, 1f);
+        fuenteVida.draw(RecursosGlobales.batch, texto, posX - offset, posY - offset);
+        fuenteVida.draw(RecursosGlobales.batch, texto, posX + offset, posY - offset);
+        fuenteVida.draw(RecursosGlobales.batch, texto, posX - offset, posY + offset);
+        fuenteVida.draw(RecursosGlobales.batch, texto, posX + offset, posY + offset);
+
+        fuenteVida.setColor(1, 1, 1, 1f);
+        fuenteVida.draw(RecursosGlobales.batch, texto, posX, posY);
     }
 
-    public void mostrarContador(SpriteBatch batch, float contador, Camara camara) {
-        String texto = String.format("Tiempo restante: %.2f s", contador);
-        layout.setText(fuente, texto);
+
+    public void mostrarContador(float contador, Camara camara) {
+        String texto = String.format("%.2f", contador);
+        layout.setText(fuenteContador, texto);
 
         OrthographicCamera camera = camara.getCamera();
         Viewport viewport = camara.getViewport();
@@ -42,9 +57,9 @@ public class Hud {
         float posX = camX + offsetX;
         float posY = camY + offsetY;
 
-        fuente.draw(batch, texto, posX, posY);
+        fuenteContador.draw(RecursosGlobales.batch, texto, posX, posY);
 
     }
 
-    public void dispose() { fuente.dispose(); }
+    public void dispose() {  }
 }